+++ drizzle/migrations/0047_black_vermin.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "Users" ALTER COLUMN "is_indexable" SET DEFAULT true;