+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Please see [Database Changes](#database-changes) and [New Configuration](#new-co
 -   [x] ✏️ `<div>` and `<span>` tags are now allowed in Markdown.
 -   [x] 🔥 Removed nonstandard `/api/v1/accounts/id` endpoint (the same functionality was already possible with other endpoints).
 -   [x] ✨️ Implemented rate limiting support for API endpoints.
+-   [x] 🔒 Implemented `is_indexable` and `is_hiding_collections` fields to the [**Accounts API**](https://docs.joinmastodon.org/methods/accounts/#update_credentials).
 
 ### CLI